backend/requirements.txt
# flask==2.2.3
# flask-cors==3.0.10
# psycopg2-binary==2.9.5
# redis==4.5.1
# pika==1.3.1
# sqlalchemy==2.0.5
# gunicorn==20.1.0
# # Removed werkzeug — let Flask pull the right version
flask==2.2.3
werkzeug==2.2.3
flask-cors==3.0.10
psycopg2-binary==2.9.5
redis==4.5.1
pika==1.3.1
sqlalchemy==2.0.5
gunicorn==20.1.0
